+++ lib/massive_record/rails/railtie.rb
@@ -3,12 +3,6 @@ module MassiveRecord
     class Railtie < ::Rails::Railtie
       config.massive_record = ActiveSupport::OrderedOptions.new
 
-      initializer "massive_record.set_configs" do |app|
-        ActiveSupport.on_load(:massive_record) do
-          app.config.massive_record.each { |k,v| send("#{k}=", v) }
-        end
-      end
-
       initializer "massive_record.logger" do
         MassiveRecord::ORM::Base.logger = ::Rails.logger
       end
@@ -28,10 +22,23 @@ module MassiveRecord
         end
       end
 
+      # Insert IdentityMap's middleware if enabled
+      initializer "massive_record.identity_map" do |app|
+        if config.massive_record.delete(:identity_map)
+          config.app_middleware.insert_after "::ActionDispatch::Callbacks", "MassiveRecord::ORM::IdentityMap::Middleware"
+        end
+      end
+
       initializer "massive_record.time_zone_awareness" do
         ActiveSupport.on_load(:massive_record) do
           self.time_zone_aware_attributes = true
           self.default_timezone = :utc
+        end
+      end
+
+      initializer "massive_record.set_configs" do |app|
+        ActiveSupport.on_load(:massive_record) do
+          app.config.massive_record.each { |k,v| send("#{k}=", v) }
         end
       end